Module amazon_sp_fulfillment_outbound::apis::fba_outbound_api

Functions

cancel_fulfillment_order

Requests that Amazon stop attempting to fulfill the fulfillment order indicated by the specified order identifier. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

create_fulfillment_order

Requests that Amazon ship items from the seller’s inventory in Amazon’s fulfillment network to a destination address. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

create_fulfillment_return

Creates a fulfillment return. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_feature_inventory

Returns a list of inventory items that are eligible for the fulfillment feature you specify. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_feature_sku

Returns the number of items with the sellerSKU you specify that can have orders fulfilled using the specified feature. Note that if the sellerSKU isn’t eligible, the response will contain an empty skuInfo object. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_features

Returns a list of features available for Multi-Channel Fulfillment orders in the marketplace you specify, and whether the seller for which you made the call is enrolled for each feature. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_fulfillment_order

Returns the fulfillment order indicated by the specified order identifier. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_fulfillment_preview

Returns a list of fulfillment order previews based on shipping criteria that you specify. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

get_package_tracking_details

Returns delivery tracking information for a package in an outbound shipment for a Multi-Channel Fulfillment order. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

list_all_fulfillment_orders

Returns a list of fulfillment orders fulfilled after (or at) a specified date-time, or indicated by the next token parameter. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

list_return_reason_codes

Returns a list of return reason codes for a seller SKU in a given marketplace. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.

update_fulfillment_order

Updates and/or requests shipment for a fulfillment order with an order hold on it. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 2 | 30 | For more information, see "Usage Plans and Rate Limits" in the Selling Partner API documentation.