Struct amazon_sp_fulfillment_outbound::models::create_fulfillment_order_request::CreateFulfillmentOrderRequest

pub struct CreateFulfillmentOrderRequest {

    pub marketplace_id: Option<String>,
    pub seller_fulfillment_order_id: String,
    pub displayable_order_id: String,
    pub displayable_order_date: String,
    pub displayable_order_comment: String,
    pub shipping_speed_category: ShippingSpeedCategory,
    pub delivery_window: Option<Box<DeliveryWindow>>,
    pub destination_address: Box<Address>,
    pub fulfillment_action: Option<FulfillmentAction>,
    pub fulfillment_policy: Option<FulfillmentPolicy>,
    pub cod_settings: Option<Box<CodSettings>>,
    pub ship_from_country_code: Option<String>,
    pub notification_emails: Option<Vec<String>>,
    pub feature_constraints: Option<Vec<FeatureSettings>>,
    pub items: Vec<CreateFulfillmentOrderItem>,

}

CreateFulfillmentOrderRequest : The request body schema for the createFulfillmentOrder operation.

Fields
marketplace_id: Option<String>

The marketplace the fulfillment order is placed against.

seller_fulfillment_order_id: String

A fulfillment order identifier that the seller creates to track their fulfillment order. The SellerFulfillmentOrderId must be unique for each fulfillment order that a seller creates. If the seller’s system already creates unique order identifiers, then these might be good values for them to use.

displayable_order_id: String

A fulfillment order identifier that the seller creates. This value displays as the order identifier in recipient-facing materials such as the outbound shipment packing slip. The value of DisplayableOrderId should match the order identifier that the seller provides to the recipient. The seller can use the SellerFulfillmentOrderId for this value or they can specify an alternate value if they want the recipient to reference an alternate order identifier. The value must be an alpha-numeric or ISO 8859-1 compliant string from one to 40 characters in length. Cannot contain two spaces in a row. Leading and trailing white space is removed.

displayable_order_date: Stringdisplayable_order_comment: String

Order-specific text that appears in recipient-facing materials such as the outbound shipment packing slip.

shipping_speed_category: ShippingSpeedCategorydelivery_window: Option<Box<DeliveryWindow>>destination_address: Box<Address>fulfillment_action: Option<FulfillmentAction>fulfillment_policy: Option<FulfillmentPolicy>cod_settings: Option<Box<CodSettings>>ship_from_country_code: Option<String>

The two-character country code for the country from which the fulfillment order ships. Must be in ISO 3166-1 alpha-2 format.

notification_emails: Option<Vec<String>>

A list of email addresses that the seller provides that are used by Amazon to send ship-complete notifications to recipients on behalf of the seller.

feature_constraints: Option<Vec<FeatureSettings>>

A list of features and their fulfillment policies to apply to the order.

items: Vec<CreateFulfillmentOrderItem>

An array of item information for creating a fulfillment order.

Implementations

impl CreateFulfillmentOrderRequest

pub fn new(
seller_fulfillment_order_id: String,
displayable_order_id: String,
displayable_order_date: String,
displayable_order_comment: String,
shipping_speed_category: ShippingSpeedCategory,
destination_address: Address,
items: Vec<CreateFulfillmentOrderItem>
) -> CreateFulfillmentOrderRequest

The request body schema for the createFulfillmentOrder operation.